refactor: ListingView unit summaries to helpers
