+++ ui-components/src/locales/general.json
@@ -1615,6 +1615,7 @@
     "WY": "Wyoming"
   },
   "t": {
+    "&": "&",
     "areYouSure": "Are you sure?",
     "addNotes": "Add notes",
     "at": "at",
